Refactor: Move all fetch logic into `RequestUtils`
This is an imperfect solution because the `make_request` ->
`make_request_with_headers` flow is inelegant. We should have a unitary
`make_request` that makes use of keyword options like `body`, `session`,
and `headers` and infers from those which type of request it should
make.

In the short term this is still better than what we were doing before,
which was federating fetch logic across multiple modules. This way we at
least encapsulate all the `HTTPoison` references in one place. This will
help later on by making these files easier to test.